+++ projects/ngx-feature-toggle/src/lib/ngx-feature-toggle.module.ts
@@ -2,11 +2,10 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FeatureToggleProviderComponent } from './ngx-feature-toggle-provider.component';
 import { FeatureToggleDirective } from './ngx-feature-toggle.directive';
-import { FeatureToggleWhenDisabledDirective } from './ngx-feature-toggle-when-disabled.directive';
 
 @NgModule({
-  declarations: [FeatureToggleProviderComponent, FeatureToggleDirective, FeatureToggleWhenDisabledDirective],
-  exports: [FeatureToggleProviderComponent, FeatureToggleDirective, FeatureToggleWhenDisabledDirective],
+  declarations: [FeatureToggleProviderComponent, FeatureToggleDirective],
+  exports: [FeatureToggleProviderComponent, FeatureToggleDirective],
   imports: [CommonModule],
 })
-export class FeatureToggleModule { }
+export class FeatureToggleModule {}