+++ projects/ngx-feature-toggle/src/public-api.ts
@@ -3,7 +3,6 @@
  */
 
 export * from './lib/ngx-feature-toggle.directive';
-export * from './lib/ngx-feature-toggle-when-disabled.directive';
 export * from './lib/ngx-feature-toggle-provider.component';
 export * from './lib/ngx-feature-toggle.module';
 export * from './lib/ngx-feature-toggle-can-activate-guard.router';